alignment: don't remove alignment on dispose()
In GcsvWindow's dispose, when unreffing the GcsvAlignment, when the
alignment was removed it changed the buffer content, so the
buffer-changed signal was emitted in GcsvWindow. But since GcsvWindow
was already in dispose(), there was a crash. The signal is disconnected
only when chaining-up the dispose(), but usually the chain-up in
dispose() is at the end, it would have been strange to chain-up first.

So it's simpler to not remove the alignment when destroying a
GcsvAlignment. Instead a function has been added to remove the alignment
explicitly.