align: store columns lengths
The goal is to align the columns asynchronously, to not block the UI. So
we need to store the columns lengths to avoid the need to re-scan the
whole buffer for each chunk to align.

Another way would have been to compute the column length by looking at
an already aligned line. But I think it's less robust and more prone to
errors.

Also, use guint's instead of gint's where appropriate.