align: block/unblock all "modified-changed" signal handlers
We don't want to send a plethora of signals when doing the alignment.
The modified bit of the buffer is restored after doing an alignment, so
we can block all signal handlers. But GObject doesn't have a convenient
way to block all signal handlers, so implement that ourselves in the
utils functions.