Ask before closing an unsaved file
The document name is needed, so store it in the object.

The close confirmation dialog doesn't have a Save button because it's a
bit more complicated to implement, and it would be re-inventing the
wheel (it's already implement in gedit for example, and hopefully that
code in gedit will be re-usable in a not-too-distant future).