delimiter-chooser: use a simple combo box
There will be a separate GtkEntry with an "Other" item in the combo.
Having the entry combined with the combo box is a bit strange.

Maybe in the future there will be a redesigned combo box widget in GTK+
with a popover. In that case it'll maybe be better to use it.