Rename GcsvDelimiterChooser -> GcsvPropertiesChooser
Since the class now contains another property (the title line).

GcsvPropertiesChooser seems a good name. "Settings" can be confusing
since in some apps it's for the GSettings. "Preferences" is not a good
name either, since it's not really the "user preference", it's more the
"DSV file properties". A more complete name would be
GcsvDsvPropertiesChooser, or even GcsvDsvFilePropertiesChooser. Or
"View" instead of "Chooser". I think GcsvPropertiesChooser is clear
enough...

It is possible in the future to create GcsvBuffer, a subclass of
GtkSourceBuffer to add the two (or more) properties: the delimiter and the
title line. So the PropertiesChooser would configure directly the
GcsvBuffer, and the GcsvAlignment would read and listen to the
properties of GcsvBuffer, instead of duplicating the properties in
several classes.