window: show_all when the window is fully constructed
There was a problem with the app-menu. The app-menu was not shown, there
was only the Quit action.
It turned out that the problem is that gtk_widget_show_all() must not be
called in the init() function, we must wait that the window is fully
constructed. I don't know why, but it works.