Move to SourceForge
GitHub is nice as a Git repo, but beyond that it's not that great. For
example for the downloads, GitHub creates itself tarballs for each tag
in Git. But those tarballs must not be used. With the Autotools, the
tarballs must be created with the 'make distcheck' command, which
generates additional files, for example the configure shell script.

On SourceForge, we upload the files that we want. There is a consistent
summary for each software on the forge: a short description, a
screenshot, a long description, a list of features, a list of tags, etc.
The features of SourceForge are nicely integrated. And gCSVedit will
maybe have more visibility. With the tags "text editor" and "CSV", it's
easy to find gCSVedit.